Fix x86 rom version build when CONFIG_X86_RESET_VECTOR is on
Currently U-Boot on x86 only supports chromebook-x86 which relies
on the first stage bootloader 'coreboot' to handle things like DRAM
initialization and processor/chipset specific low level stuff. In
this case, CONFIG_X86_RESET_VECTOR is not defined thus the reset
vector and 16-bit real mode codes are not included in the U-Boot.

When turning on the CONFIG_X86_RESET_VECTOR, the linking process
acutally misses the resetvec.o and start16.o so it cannot generate
the rom version of U-Boot. The top level Makefile is updated to
pull them into the final linking process.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>