x86: save FspInit returned results to global data
FspInit() returns HobList pointer and SharedData pointer to U-Boot
and we save them to the global data. Two members are added to the
end of the arch_global_data which is also the end of the global
data structure.

Note the global data space on the stack must be cleared to zero,
otherwise random crash will be observed in later board_init_f()
call since the stack space was already polluted in the FspInit()
ContinuationFunc routine.

Signed-off-by: Bin Meng <bmeng.cn@gmail.com>